Changes pipeline.py. It will not erase the existing index.html. It will append to that file
git-svn-id: https://pymorph.googlecode.com/svn@7 2dd86497-8ce7-14b7-f524-3852097d2217

+++ serial_pipeline/trunk/pipeline.py
@@ -25,10 +25,13 @@ def main():
     size = c.size/2
     threshold = c.threshold
     thresh_area = c.thresh_area
-    indexfile = open('index.html', 'w')
-    indexfile.writelines(['<HTML>\n<BODY>\n'])
-    indexfile.writelines(['</BODY></HTML>'])
-    indexfile.close()
+    if exists('index.html'):
+        pass
+    else:
+        indexfile = open('index.html', 'w')
+        indexfile.writelines(['<HTML>\n<BODY>\n'])
+        indexfile.writelines(['</BODY></HTML>'])
+        indexfile.close()
     try:
         img = pyfits.open(imagefile)
         image = img[0].data
